rename and remove old files
