+++ miniwdl_viz/miniwdl_parser.py
@@ -1,167 +1,51 @@
+import argparse
+import yaml
+import json
+from pathlib import Path
+
 import WDL
-import re
-
-SELECT_ALL = "select_all"
-SELECT_FIRST = "select_first"
-DEFINED = "defined"
-FLATTEN = "flatten"
-LENGTH = "length"
-FLOOR = "floor"
-DIV = "_div"
-AT = "_at"
-READ_LINES = "read_lines"
-GLOB = "glob"
-
-function_map = {"_eqeq": " == ", "_neq": " != ", "_land": " && ", "_lte": " <= "}
 
 
 class MiniWDLParser:
-    """
-
-    TODO: Support function declarations (eg. select_first)
-    TODO: add edges in the items
-    """
+    """Parses a miniwdl file into inputs, outputs, nodes, and edges"""
 
     def __init__(self, wdl_doc):
         self.wdl_doc = wdl_doc
+        self.workflow_name = ""
+        self.nodes = []
         self.edges = []
-        self.declarations = {}
+        self.inputs = {}
+        self.outputs = {}
 
-    def parse_workflow_task(self, task):
-        """parses a task in a workflow
-        a task can be one of either:
-        Call,
-        Conditional,
-        Declaration
-        returns a list of parsed tasks
-        """
-        if isinstance(task, WDL.Tree.Call):
-            return self.read_call_task(task)
-        elif isinstance(task, WDL.Tree.Conditional) or isinstance(
-            task, WDL.Tree.Scatter
-        ):
-            return self.read_conditional_task(task)
-        elif isinstance(task, WDL.Tree.Decl):
-            return self.read_declaration_task(task)
+    def parse_inputs(self, inputs):
+        if inputs is None: 
+            return
+        for input in inputs:
+            self.inputs[input.name] = str(input.type)
+
+    def parse_outputs(self, outputs):
+        if outputs is None:
+            return
+        for output in outputs:
+            self.outputs[output.name] = str(output.type)
+
+    def get_referee_name(self, referee):
+        if isinstance(referee, WDL.Gather):
+            return referee.final_referee.workflow_node_id
         else:
-            breakpoint()
-            raise Exception(f"Unsupported workflow task type")
+            return referee.workflow_node_id
 
-    ###################
-    ### Read Tasks ####
-    ###################
-
-    def read_conditional_task(self, conditional):
-        """parses a conditional task
-        a conditional task is an "if" condition that contains
-        tasks to run
-        """
-        tasks = []
-        cond_string = self.parse_conditional_item(conditional.expr)
-        for item in conditional.body:
-            ## TODO: fix issue of nested if statements
-            parsed_items = self.parse_workflow_task(item)
-            for item in parsed_items:
-                self.edges.append(
-                    {
-                        "node_from": cond_string,
-                        "node_to": item["name"],
-                        "workflow_name": "true",
-                        "task_name": "true",
-                        "edge_type": "cond",
-                    }
-                )
-            tasks.extend(parsed_items)
-
-        return tasks
-
-    def read_declaration_task(self, declaration):
-        decl = {"name": declaration.name, "inputs": [], "type": type(declaration)}
-
-        expression = declaration.expr
-        parsed_expression = []
-        if expression is None:
-            """If input declaration is outside of the input section
-            (not recommended)
-            """
-            return [decl]
-        elif type(expression) == WDL.Expr.IfThenElse:
-            parsed_expression.extend(self.parse_input_item(expression.consequent))
-            parsed_expression.extend(self.parse_input_item(expression.alternative))
-            # declaration name is the name of the variable being declared
-            self.declarations[declaration.name] = "if_then_else"
-            decl["inputs"] = parsed_expression
-        else:
-            if hasattr(expression, "function_name"):
-                parsed_expression = self.parse_input_item(expression)
-                operator = function_map.get(expression.function_name, "")
-                expression_name = operator.join(
-                    [
-                        re.sub('\(|\)|"', "_", str(argument))
-                        for argument in expression.arguments
-                    ]
-                )
-                self.declarations[declaration.name] = str(expression_name)
-
-        for reference_string in parsed_expression:
-            output_from, output_var, intype = self.create_node_variables(
-                reference_string=reference_string, intype="decl_to"
-            )
-
-            self.edges.append(
-                {
-                    "node_from": output_from,
-                    "node_to": self.declarations[declaration.name],
-                    "workflow_name": output_var,
-                    "task_name": output_var,
-                    "edge_type": intype,
-                }
-            )
-        return [decl]
-
-    def read_call_task(self, call):
-        task = {"name": call.name, "inputs": [], "type": type(call)}
-        # For each input to the call
-        for short_name, reference in call.inputs.items():
-            # eg. short_name = reads1_fastq
-            # eg. reference = fastp_qc.fastp1_fastq
-
-            reference_strings = self.parse_input_item(reference)
-            for reference_string in reference_strings:
-                output_from, output_var, intype = self.create_node_variables(
-                    reference_string=reference_string, intype="input"
-                )
-                self.edges.append(
-                    {
-                        "node_from": output_from,
-                        "node_to": call.name,
-                        "workflow_name": output_var,
-                        "task_name": short_name,
-                        "edge_type": intype,
-                    }
-                )
-                # add to files list
-                key = ".".join([output_from, output_var])
-                task["inputs"].append(key)
-        return [task]
-
-    ###################
-    ### Parse Items ###
-    ###################
-
-    def parse_input_item(self, reference):
-        if isinstance(reference, WDL.Expr.Get):
-            return self.parse_get_expression(reference)
-        elif isinstance(reference, WDL.Expr.Apply):
-            return self.parse_apply_expression(reference)
-        elif isinstance(reference, WDL.Expr.Array):
-            return self.parse_array_expression(reference)
-        elif isinstance(reference, WDL.Expr.IfThenElse):
-            return self.parse_if_then_else_expression(reference)
+    def parse_input_item(self, input_item, node_ids):
+        if isinstance(input_item, WDL.Expr.Ident):
+            # If referee is not a defined node_id, probably comes from workflow_input
+            if (ref_name := self.get_referee_name(input_item.referee)) in node_ids:
+                ref = ref_name
+            else:
+                ref = "WorkflowInput"
+            return [{"name": f"{input_item.name}", "ref": f"{ref}", "type": "input"}]
         elif isinstance(
-            reference,
+            input_item,
             (
-                WDL.Expr.Ident,
                 WDL.Expr.String,
                 WDL.Expr.Int,
                 WDL.Expr.Float,
@@ -169,76 +53,143 @@ class MiniWDLParser:
                 WDL.Expr.Null,
             ),
         ):
-            # TODO: add hard-coded constants to edges
-            return [str(reference)]  # ignore hard-coded constants
-        breakpoint()
+            return [
+                {"name": f"{input_item}", "ref": "HardcodedVariable", "type": "input"}
+            ]
+        elif isinstance(input_item, WDL.Expr.Base):
+            names = []
+            for child_input_item in input_item.children:
+                names.extend(self.parse_input_item(child_input_item, node_ids))
+            return names
 
-        raise Exception(f"Unsupported input item: {reference}")
-
-    def parse_if_then_else_expression(self, if_else_expression):
-        parsed_expression = []
-        parsed_expression.extend(self.parse_input_item(if_else_expression.consequent))
-        parsed_expression.extend(self.parse_input_item(if_else_expression.alternative))
-        return parsed_expression
-
-    def parse_get_expression(self, get_exp):
-        expression = get_exp.expr
-        if isinstance(expression, WDL.Expr.Ident):
-            return [str(expression.name)]
-        else:
-            raise Exception(f"Unsupported get expression")
-
-    def parse_apply_expression(self, apply_exp):
-        function_name = str(apply_exp.function_name)
-        if function_name == DEFINED:
-            return []
-        else:
-            return self.parse_array_expression(apply_exp.arguments[0])
-
-    def parse_array_expression(self, array_exp):
-        items = []
-        for item in array_exp.children:
-            items.extend(self.parse_input_item(item))
-        return items
-
-    def parse_conditional_item(self, expr):
-        # TODO: add function_names
-        # Try to add function to `parse_input_item`
-        if type(expr) == WDL.Expr.Get:
-            return ",".join(self.parse_input_item(expr))
-
-        operator = function_map.get(expr.function_name, "")
-
-        return operator.join(
-            [re.sub('\(|\)|"', "_", str(argument)) for argument in expr.arguments]
+    def add_node(self, nodes, parent, id, name, type):
+        section = (
+            parent.name if isinstance(parent, WDL.Workflow) else parent.workflow_node_id
         )
+        nodes.append({"id": id, "name": name, "section": section, "type": type})
 
-    ######################
-    ### Misc Functions ###
-    ######################
+    def add_edge(self, nodes, edges, input, node_id, task_name=None):
+        input_param_list = self.parse_input_item(
+            input, [node["id"] for node in nodes]
+        )
+        for input_param in input_param_list:
+            if not task_name:
+                task_name = self.remove_prefix(
+                        input_param["name"],
+                        self.remove_prefix(input_param["ref"], "call-") + ".",
+                )
 
-    def create_node_variables(self, reference_string, intype="input"):
-        if reference_string.startswith('"') and reference_string.endswith('"'):
-            return "HardcodedInput", reference_string, intype
-
-        expression_components = reference_string.split(".")
-
-        if (
-            len(expression_components) > 1
-        ):  # The file comes from another task in this stage
-            return expression_components[0], expression_components[1], intype
-        elif self.declarations.get(
-            reference_string, None
-        ):  # If the variable is a declaration
-            # breakpoint()
-            return (
-                self.declarations[reference_string],
-                str(reference_string),
-                "decl_from",
+            edges.append(
+                {
+                    "node_from": input_param["ref"],
+                    "node_to": node_id,
+                    "workflow_ref_name": input_param["name"],
+                    "task_ref_name": task_name,
+                    "edge_type": input_param["type"],
+                }
             )
-        else:
-            return "WorkflowInput", reference_string, intype
+
+    def parse_workflow(self, wdl_doc, nodes, edges):
+        """Parse WDL workflow. Handles WDL Workflow, Call, Decl, and WorkflowSection (if/scatter) nodes"""
+        if isinstance(wdl_doc, WDL.Workflow):
+            """If workflow, parse the inputs, outputs and recursively parse each call"""
+            if not self.workflow_name:
+                self.workflow_name = wdl_doc.name
+            self.parse_inputs(wdl_doc.inputs)
+            for task in wdl_doc.body:
+                self.parse_workflow(task, nodes, edges)
+            self.parse_outputs(wdl_doc.outputs)
+
+        elif isinstance(wdl_doc, WDL.Call):
+            """If Call, add the node and parse each input item to determine the type and referee"""
+            self.add_node(
+                nodes=nodes,
+                parent=wdl_doc.parent,
+                id=wdl_doc.workflow_node_id,
+                name=wdl_doc.name,
+                type="call",
+            )
+            for task_name, workflow_input in wdl_doc.inputs.items():
+                self.add_edge(nodes, edges, workflow_input, wdl_doc.workflow_node_id, task_name)
+
+        elif isinstance(wdl_doc, WDL.WorkflowSection):
+            """If WorkflowSection (if/scatter), add node and recurse on each call in the section"""
+            self.add_node(
+                nodes=nodes,
+                parent=wdl_doc.parent,
+                id=wdl_doc.workflow_node_id,
+                name=str(str(wdl_doc.expr)),
+                type="workflow_section",
+            )
+            self.add_edge(nodes, edges, wdl_doc.expr, wdl_doc.workflow_node_id)
+            for task in wdl_doc.body:
+                self.parse_workflow(task, nodes, edges)
+
+        elif isinstance(wdl_doc, WDL.Decl):
+            """If Decl, add node, then parse edges to add the type and refree"""
+            if wdl_doc.expr:
+                self.add_node(
+                    nodes=nodes,
+                    parent=wdl_doc.parent,
+                    id=wdl_doc.workflow_node_id,
+                    name=wdl_doc.name,
+                    type="decl",
+                )
+                self.add_edge(nodes, edges, wdl_doc.expr, wdl_doc.workflow_node_id)
+
+    @staticmethod
+    def remove_prefix(string, prefix):
+        if string.startswith(prefix):
+            return string[len(prefix) :]
+        return string
 
     def parse(self):
-        for task in self.wdl_doc.workflow.body:
-            self.parse_workflow_task(task)
+        self.parse_workflow(self.wdl_doc.workflow, self.nodes, self.edges)
+
+    def to_dict(self):
+        return {
+            "inputs": self.inputs,
+            "outputs": self.outputs,
+            "nodes": self.nodes,
+            "edges": self.edges
+        }
+
+def write_output(filename, out_type, dict):
+    with open(f"{filename}.{out_type}", "w") as f:
+        if out_type == "yaml":
+            yaml.dump(
+                dict, 
+                f,
+                sort_keys=False
+            )
+        elif out_type == "json":
+            json.dump(
+                dict,
+                f,
+                indent=4
+            )
+
+def main():
+    arg_parser = argparse.ArgumentParser(
+        prog="MiniWDLParser",
+        description=MiniWDLParser.__doc__,
+    )
+    arg_parser.add_argument("input_wdl")
+    arg_parser.add_argument("-d", "--output-dir")
+    arg_parser.add_argument("-o", "--output-file")
+    arg_parser.add_argument("-j", "--json-output", action='store_true', default=False)
+
+    args = arg_parser.parse_args()
+
+    doc = WDL.load(uri=args.input_wdl)
+    parser = MiniWDLParser(doc)
+    parser.parse()
+
+    out_type = "json" if args.json_output else "yaml"
+    filename = args.output_file if args.output_file else Path(args.input_wdl).stem 
+    filepath = str(Path(args.output_dir)/filename) if args.output_dir else filename
+    write_output(filepath, out_type, parser.to_dict())
+    
+
+if __name__ == "__main__":
+    main()
